requirements.txt
apipkg==1.4
beautifulsoup4==4.4.1
cffi==1.5.2
cryptography==1.3.1
decorator==4.1.2
enum34==1.1.3
execnet==1.5.0
Flask==0.10.1
FormEncode==1.3.1
html2text==2016.9.19
idna==2.1
IMDbPY==5.1.1
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
keen==0.5.1
lxml==3.8.0
MarkupSafe==0.23
ndg-httpsclient==0.4.0
pbr==3.1.1
py==1.4.34
pyasn1==0.1.9
pycparser==2.14
pycryptodome==3.4.7
PyDispatcher==2.0.5
pyOpenSSL==16.0.0
pytest==3.2.3
pytest-forked==0.2
pytest-sugar==0.9.0
pytest-xdist==1.20.1
requests==2.9.1
requests-cache==0.4.13
six==1.10.0
SQLAlchemy==1.1.13
sqlalchemy-migrate==0.11.0
SQLObject==3.4.0
sqlparse==0.2.3
Tempita==0.5.2
termcolor==1.1.0
Werkzeug==0.11.8
wikipedia==1.4.0
Pillow==5.0.0
